requirements.txt
Flask
pillow
matplotlib
numpy
pandas
opencv-python
gunicorn
tensorflow-cpu
